Breaking(All): Separation of readIn/readFiles and tests
HMJ -> {HMJ, core, files, stdin} with updated tests and some changes

BREAKING CHANGE: Functions not available from `hashmyjs`

re #40
