+++ cc.txt
@@ -1,0 +1,60 @@
+
+> hashmyjs@2.0.0 cc:all /mnt/c/Users/max/Projects/hashmyjs
+> npm run cc *.js src/*.js
+
+
+> hashmyjs@2.0.0 cc /mnt/c/Users/max/Projects/hashmyjs
+> cr -c ./config/.complexrc "index.js" "src/clr.js" "src/hashmyjs.js" "src/utils.js"
+
+# Complexity report, 2018-10-16
+
+* Mean per-function logical LOC: 15.75
+* Mean per-function parameter count: 0
+* Mean per-function cyclomatic complexity: 1
+* Mean per-function Halstead effort: 1588.0288722490827
+* Mean per-module maintainability index: 61.15591798083482
+* First-order density: 0%
+* Change cost: 25%
+* Core size: 0%
+
+## index.js
+
+* Physical LOC: 26
+* Logical LOC: 7
+* Mean parameter count: 0
+* Cyclomatic complexity: 1
+* Cyclomatic complexity density: 14.285714285714285%
+* Maintainability index: 67.33062891804673
+* Dependency count: 5
+
+## src/clr.js
+
+* Physical LOC: 9
+* Logical LOC: 8
+* Mean parameter count: 0
+* Cyclomatic complexity: 1
+* Cyclomatic complexity density: 12.5%
+* Maintainability index: 68.53831737650921
+* Dependency count: 0
+
+## src/hashmyjs.js
+
+* Physical LOC: 288
+* Logical LOC: 26
+* Mean parameter count: 0
+* Cyclomatic complexity: 1
+* Cyclomatic complexity density: 3.8461538461538463%
+* Maintainability index: 53.48892060158285
+* Dependency count: 6
+
+## src/utils.js
+
+* Physical LOC: 107
+* Logical LOC: 22
+* Mean parameter count: 0
+* Cyclomatic complexity: 1
+* Cyclomatic complexity density: 4.545454545454546%
+* Maintainability index: 55.26580502720052
+* Dependency count: 2
+
+
