+++ src/hashmyjs.js
@@ -6,236 +6,10 @@
  */
 /* eslint-env es6, node */
 
-/**
- * Configuration
- * @typedef {Object} Config
- * @property {boolean} [prettify=false] Prettify the output
- * @property {string} [outputDest=OUTPUT_DEST] Output destination (stdout, var, <i>filename</i>)
- * @property {string} [outputFormat=OUTPUT_FORMAT] Output format (text, json, csv)
- */
-
-const sjcl = require('sjcl'),
-  readline = require('readline'),
-  fs = require('fs'),
-  clr = require('colors/safe');
-const { IoError, info, out, EOF } = require('./utils');
-clr.setTheme(require('./clr'));
-
-/**
- * @description Default output destination.
- * @summary stdout
- * @constant
- * @type {string}
- */
-const OUTPUT_DEST = 'stdout';
-
-/**
- * @description Default output format.
- * @summary text
- * @constant
- * @type {string}
- */
-const OUTPUT_FORMAT = 'text';
-
-/**
- * @description Generate a base64-encoded SHA-256 hash for a given data.<br>
- * <em>Objects can be passed to this function but will generate the <strong>same</strong> hash regardless
- * of its content</em>.
- * @param {(string|number|Array|Date)} data Data to encode
- * @return {string} Base64 encoded SHA-256 hash
- * @public
- * @example hash('Lorem Ipsum dolore sit amet'); //returns 'sha256-7almix3trlcKWVAN+fhV/Bzbx4BixTwzjYpZDWUxctM='
- */
-const hash = (data) => {
-  let hashed = sjcl.hash.sha256.hash(data);
-  return `sha256-${sjcl.codec.base64.fromBits(hashed)}`;
-};
-
-/**
- * @description Write to a file (while prettifying the content).
- * @param {string} filename File name
- * @param {string[]} data Lines to write to the file
- * @param {string} [outputFormat=OUTPUT_FORMAT] Format of the output
- * @throws {Error} No filename specified
- * @throws {IoError} Writing error
- * @public
- * @example writeToFile('output.txt', ['Lorem ipsum dolore sit amet']);
- * @example <caption>With a specified format:</caption>
- * writeToFile('output.json', [{key: 'val'}], 'json');
- * writeToFile('output.csv', ['0,john,doe', '1,lorem,ipsum'], 'csv');
- */
-const writeToFile = (filename, data, outputFormat = OUTPUT_FORMAT) => {
-  if (!filename) throw new Error(`No filename specified to be written to with data=${data}`);
-  filename = ('' + filename).trim();
-  fs.writeFile(filename, '', (err) => {
-    if (err) throw new Error(`Couldn't write in ${filename}`);
-    let writer = fs.createWriteStream(filename, {
-      flags: 'a'
-    });
-    data.forEach((line) => writer.write(`${prettifyOutput(line, outputFormat)}\n`));
-    info(`Successfully written the result to ${filename}`);
-  });
-};
-
-/**
- * @description Prettify the ouptut according to the format used or keep it as it is.
- * @param {(string|Object|string[])} output Output
- * @param {string} [format=OUTPUT_FORMAT] Format of the output (json, csv, ...)
- * @return {string} Prettified output
- * @public
- * @example prettifyOutput('some unchanged text'); //returns 'some unchanged text'
- * @example <caption>With a specified format:</caption>
- * prettifyOutput({key: 'val'}, 'json'); //returns {<br>  "key": "val"<br>}
- * prettifyOutput('0,john,doe', 'csv'); //returns '0, joh, doe'
- */
-const prettifyOutput = (output, format = OUTPUT_FORMAT) => {
-  switch (format) {
-  case 'json':
-    return JSON.stringify(output, null, 2);
-  case 'csv':
-    const re = /,(?=[^\s])/g;
-    return Array.isArray(output) ? output.map(item => item.replace(re, ', ')) : output.replace(re, ', ');
-  default:
-    return output;
-  }
-};
-
-/**
- * @description Handle JSON data.
- * @param {Object|string|Array} json JSON data
- * @param {boolean} [prettify=false] Make it pretty
- * @returns {Object|string|Array} (Stringified) JSON data
- */
-const jsonHandler = (json, prettify = false) => {
-  return prettify ? JSON.stringify(json, null, 2) : json;
-};
-
-/**
- * @description Handle CSV data.
- * @param {string} lhs CSV data on the left hand side
- * @param {string} rhs CSV data on the right hand side
- * @param {boolean} [prettify=false] Make it pretty
- * @returns {string} CSV data
- */
-const csvHandler = (lhs, rhs, prettify = false) => {
-  return prettify ? `${lhs}, ${rhs}` : `${lhs},${rhs}`;
-};
-
-/**
- * @description Scan an input and output it's integrity hash.
- * @param {(string|string[])} input Input to hash (e.g. JS code)
- * @param {boolean} [noOutput=false] Don't output the result to the terminal but return the hash
- * @return {(void|string)} Hashed data or nothing
- * @throws {Error} Hashing or input error
- * @public
- * @example scanInput('Lorem ipsum dolore sit amet'); //logs 'sha256-HBQ/am1i8gw1bl8qJDhm0naAsChqeYsBEiCWTRLEaE8='
- * scanInput(['Lorem ipsum dolore sit amet']); //logs the same as above
- * @example <caption>Without output disabled</caption>
- * scanInput('Lorem ipsum dolore sit amet', true); //returns 'sha256-HBQ/am1i8gw1bl8qJDhm0naAsChqeYsBEiCWTRLEaE8='
- * scanInput(['Lorem ipsum dolore sit amet']); //returns the same as above
- */
-const scanInput = (input, noOutput = false) => {
-  let data = Array.isArray(input) ? input.join('\n') : input;
-  if (!input) throw new Error('scanInput didn\'t received any input');
-  let digest = hash(data);
-  if (typeof digest !== 'string' && 'message' in digest) throw new Error(digest.message);
-  if (noOutput) return digest;
-  out(`${digest}`);
-};
-
-/**
- * @description Synchronously read files and scan them.
- * @param {string[]} [files] Array of file paths
- * @param {Config} obj Configuration.
- * @see Config
- * @return {(undefined|string[]|{...string})} Data or nothing
- * @public
- * @example <caption>Reading from the CLI</caption>
- * readFilesSync();
- * @example <caption>Reading from specific files</caption>
- * readFilesSync(['output.txt']);
- * @example <caption>... With specific configurations</caption>
- * readFilesSync(['input.json'], {prettify: true, outputFormat: 'json'}); //logs {<br>  "output.json": "sha256-iTyF6rE+vAUIIWrWaC6bWt9NwI/74kpOuk4JZl9zCMM="<br>}
- * readFilesSync(['input.csv'], {outputDest: 'output.json', outputFormat: 'json'}); //Writes the above to output.json
- */
-const readFilesSync = (files = process.argv.slice(2, process.argv.length), { prettify = false, outputDest = OUTPUT_DEST, outputFormat = OUTPUT_FORMAT } = {}) => {
-  let inputs = [],
-    res = {},
-    fileLines = [];
-
-  const TO_STDOUT = outputDest === 'stdout',
-    TO_VAR = outputDest === 'var',
-    AS_CSV = outputFormat === 'csv',
-    AS_JSON = outputFormat === 'json';
-
-  for (let i = 0, len = files.length; i < len; ++i) {
-    inputs.push(fs.readFileSync(files[i], (err) => {
-      if (err) throw new IoError(err.message);
-    }));
-
-    let data = scanInput(inputs[i], true);
-
-    res[files[i]] = data;
-    if (TO_STDOUT) {
-      if (outputFormat === 'text') out(`${i > 0 ? '\n' : ''}- ${files[i]}: ${data}`);
-      else if (AS_CSV) out(csvHandler(files[i], data, prettify));
-    } else if (!TO_VAR) {
-      if (AS_CSV) fileLines.push(csvHandler(files[i], data, prettify));
-      else fileLines.push(`- ${files[i]}: ${data}`);
-    }
-  }
-
-  if (TO_VAR) {
-    let result = [];
-    if (AS_JSON) return jsonHandler(res, prettify);
-    else if (AS_CSV) {
-      for (let file in res) result.push(csvHandler(file, res[file], prettify));
-      return result;
-    }
-    for (let file in res) result.push(res[file]);
-    return result;
-  } else if (TO_STDOUT) AS_JSON && out(JSON.stringify(res, null, prettify * 2));
-  else writeToFile(outputDest, AS_JSON ? [JSON.stringify(res, null, prettify * 2)] : fileLines);
-};
-
-/**
- * @description Read user's input from STDIN.
- * @param {Config} obj Configuration.
- * @return {(undefined|string)} Data or nothing
- * @see Config
- * @public
- * @example readIn();
- * @example <caption>With configurations</caption>
- * readIn({outputFormat: 'json'});
- * readIn({prettify: true, outputDest: 'outputFromSTDIN.txt'});
- */
-const readIn = ({ prettify = false, outputDest = OUTPUT_DEST, outputFormat = OUTPUT_FORMAT } = {}) => {
-  info('Press CTRL+D (or CMD+D or using `C` instead of `D`) to stop the STDIN reader\nType either \\$ or \\EOF in an empty line to signal an End-Of-File (this line won\'t be counted)\n');
-  let rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-      terminal: false
-    }),
-    lines = [],
-    res = null;
-
-  return new Promise((resolve, reject) => {
-    rl.on('line', (line) => {
-      if (EOF(line)) {
-        res = scanInput(lines, true);
-        let output = `- STDIN: ${res}`; //outputFormat = 'text'
-        if (outputFormat === 'json') {
-          const op = {STDIN: res};
-          output = (outputDest === 'var') ? jsonHandler(op, prettify) : JSON.stringify(op, null, prettify * 2);
-        } else if (outputFormat === 'csv') output = csvHandler('STDIN', res, prettify);
-
-        if (outputDest === 'stdout') out(output);
-        else if (outputDest !== 'var') writeToFile(outputDest, (outputFormat === 'text') ? output.split('\n') : output);
-        resolve(output);
-      } else lines.push(line);
-    });
-  })
-};
+const clr = require('colors/safe');
+  clr.setTheme(require('./clr'));
+const readIn = require('./stdin'),
+  readFilesSync = require('./files');
 
 /**
  * @description Start the hasher.
@@ -279,10 +53,6 @@ const run = (files = [], { format = 'text', input = 'any', output = 'stdout', pr
 
 module.exports = {
   run,
-  scanInput,
-  hash,
   readIn,
-  readFilesSync,
-  prettifyOutput,
-  writeToFile
+  readFilesSync
 }
