+++ test/ex0.hash
@@ -1,0 +1,1 @@
+- ./examples/ex0.js: sha256-oRWayKCc59n86sjrY/5xXsXO/t8OfjL4f2GkX0eUEII=
