+++ test/file.txt
@@ -1,0 +1,4 @@
+test
+lorem
+dolore
+sit
